src/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
email-validator
